Instantiation of celestial bodies using csv file SolarData.csv
Added SolarSystemManager.cs - goes through the SolarData.csv file to find all requested celestial bodies, then instantiates them according to prefabs that share their name, if it cannot find any it uses a "defaultprefab" which is just a sphere for now. Also added the CelestialBody.cs that allows for data to be assigned to each celestial body, which should allow for a general framework for simulation physics.