Added alternative doppler implementation and fixed plane
Added a new way of doing doppler after discussion with Oscar, new version will take the peak wavelengths from the emission spectra and shift them according to the relativistic Doppler equation. Also switched pos.y with pos.z and vel.y with vel.z which should fix the issue that the solar system is at the wrong angle for the player.